setup: Use timestamp for version in hab
This is intended to fix #18 but we are not sure if this solution works
yet as the bug in #18 is hard to reproduce. In testing so far, I hit
the bug, applied this patch and setup the cluster again and I didn't
run into the issue. However, I dropped the patch and setup the cluster
from scratch again and I wasn't able to reproduce the issue
anymore. :(

Here's hoping this iron's the bug out. But read on.

The approach in this patch is to use the current timestamp as that
increases monotonically. However, we're working with distrbuted
systems here and there is no guarantee that the system clocks of all
the involved nodes in the cluster will be in sync. As a result it is
very much possible that the timestamp from a second iteration is
lesser than what another node has already used, leading to the second
command being ignored silently since Habitat currently does not care
about informing the users if the config was applied or not. It is all
too gracious and accepts everything but silently ignores updates if
the supplied version number is older than the current version number
of the config.

That being said, we are currently using VMs on the same machine to
spin and test our cluster. This means that the clock on the VMs would
be in sync with that on the host, thus allowing us to not run into
this issue (yet). Ofcourse this may not be true if one is doing
funkier stuff with their VMs and driving the clocks out of sync
intentionally (or unintentionally).

/rant